WIP: Support for Qihoo browser
Current support for Qihoo browser isn't perfect.
For unkown reason, JavaScript codes might be evaluated in different
contexts of WebView so that global objects aren't shared.
However, this doesn't happen if we have a clean installation and try it
for the first time.

Since it is impossible to debug this browser, we simply choose not to
solve this problem in the near future.

Also, we fix version text in logs.