Implement window.close for Chromium browsers
Close #180 as finished.

`window.close` API should not be exposed to the main webpage.
Hence, we must make a stricter rule for the usage of unsafeWindow:
currently, one must use unsafeWindow to access relevant properties when
it is granted, no implicit equivalence between window and unsafeWindow
will be implied.

Especially, unsafeWindow.close != window.close when window.close is
granted.