Support script injections in frames
We use `Page.navigate` API to implement a simple frames injections, which might not work as expected when the script size is nearly 2M.

`@noframes` is also implemented, which is by default effectively true for most scripts because developers must `@grant frames` in the main page to inject into its frames.

Close #136 as completed.